base class:  Rework device address handling
To add EtherCAT aliases to the device address, the address needs to be
more opaque, and class methods deal with it correctly.

To achieve this, the `bus` and `position` keys in device data are
collapsed into a single `address` key.  This is passed around and
munged as needed by subclasses.